Upgrades and refactors
- Add more target platforms in buildscript
- Upgrade many project dependencies
- Upgrade Gradle
- Clean buildscript
- Remove unused tests
- Refactor to remove old MIDI parsing - use kmidi instead

+++ src/main/kotlin/org/wysko/midis2jam2/gui/viewmodel/SearchViewModel.kt
@@ -187,7 +187,7 @@ class SearchViewModel {
         }
 
         _matchingMidiFiles.value = searchEngine.index.filter { (_, instruments) ->
-            instruments.containsAll(selectedInstruments.value.map { generalMidiInstruments.indexOf(it) })
+            instruments.containsAll(selectedInstruments.value.map { generalMidiInstruments.indexOf(it).toByte() })
         }.keys.toList().sortedBy { it.name.lowercase() }
     }
